# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fauser Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List calendar events
      parameters:
        - name: from
          in: query
          schema:
            type: ["string", "null"]
        - name: to
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /diary:
    get:
      operationId: listDiaryEntries
      tags: [diary]
      summary: List diary entries for current user
      responses:
        "200":
          description: Entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiaryEntry"
    post:
      operationId: createDiaryEntry
      tags: [diary]
      summary: Create a diary entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiaryEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiaryEntry"
  /diary/{id}:
    patch:
      operationId: updateDiaryEntry
      tags: [diary]
      summary: Update a diary entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiaryEntryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiaryEntry"
    delete:
      operationId: deleteDiaryEntry
      tags: [diary]
      summary: Delete a diary entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /forum/threads:
    get:
      operationId: listForumThreads
      tags: [forum]
      summary: List forum threads
      parameters:
        - in: query
          name: subjectId
          schema:
            type: integer
      responses:
        "200":
          description: Threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForumThread"
    post:
      operationId: createForumThread
      tags: [forum]
      summary: Create a forum thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForumThreadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumThread"
  /forum/threads/{id}/posts:
    get:
      operationId: listForumPosts
      tags: [forum]
      summary: List posts in a thread
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForumPost"
    post:
      operationId: createForumPost
      tags: [forum]
      summary: Reply to a thread
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForumPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumPost"
...
  /justifications:
    get:
      operationId: listJustifications
      tags: [justifications]
      summary: List justifications
      parameters:
        - in: query
          name: studentId
          schema:
            type: integer
      responses:
        "200":
          description: Justifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Justification"
    post:
      operationId: createJustification
      tags: [justifications]
      summary: Submit a justification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JustificationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Justification"
  /justifications/{id}:
    patch:
      operationId: reviewJustification
      tags: [justifications]
      summary: Approve or reject a justification
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JustificationReviewInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Justification"
...
  /groups/messages:
    get:
      operationId: listGroupMessages
      tags: [groups]
      summary: List class group messages
      parameters:
        - in: query
          name: classId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassMessage"
    post:
      operationId: createGroupMessage
      tags: [groups]
      summary: Post a message to the class group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassMessageBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassMessage"
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        allDay:
          type: boolean
        type:
          type: string
          enum: [compito, esame, evento_scolastico, vacanza, altro]
        classId:
          type: ["integer", "null"]
        createdById:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - startDate
        - allDay
        - type
        - createdById
        - createdAt
    EventInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        allDay:
          type: boolean
        type:
          type: string
          enum: [compito, esame, evento_scolastico, vacanza, altro]
        classId:
          type: ["integer", "null"]
      required:
        - title
        - startDate
        - allDay
        - type
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        allDay:
          type: boolean
        type:
          type: string
          enum: [compito, esame, evento_scolastico, vacanza, altro]
...
    DiaryEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        content:
          type: string
        subjectId:
          type: integer
          nullable: true
        subjectName:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - title
        - content
        - createdAt
        - updatedAt
    DiaryEntryInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        subjectId:
          type: integer
      required:
        - title
        - content
...
    ForumThread:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        subjectName:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        classId:
          type: integer
          nullable: true
        title:
          type: string
        replyCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - subjectId
        - subjectName
        - authorId
        - authorName
        - title
        - replyCount
        - createdAt
    ForumThreadInput:
      type: object
      properties:
        subjectId:
          type: integer
        title:
          type: string
        classId:
          type: integer
      required:
        - subjectId
        - title
    ForumPost:
      type: object
      properties:
        id:
          type: integer
        threadId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - threadId
        - authorId
        - authorName
        - content
        - createdAt
    ForumPostInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...
    Justification:
      type: object
      properties:
        id:
          type: integer
        attendanceId:
          type: integer
        studentId:
          type: integer
        reason:
          type: string
        status:
          type: string
        reviewedBy:
          type: integer
          nullable: true
        reviewedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        studentName:
          type: string
        attendanceDate:
          type: string
        reviewerName:
          type: string
          nullable: true
      required:
        - id
        - attendanceId
        - studentId
        - reason
        - status
        - createdAt
    JustificationReviewInput:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
      required:
        - status
    JustificationInput:
      type: object
      properties:
        attendanceId:
          type: integer
        studentId:
          type: integer
        reason:
          type: string
      required:
        - attendanceId
        - studentId
        - reason
...
    ClassMessage:
      type: object
      properties:
        id:
          type: integer
        classId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - classId
        - senderId
        - senderName
        - content
        - createdAt
    CreateClassMessageBody:
      type: object
      properties:
        classId:
          type: integer
        content:
          type: string
      required:
        - classId
        - content